implemented money addition
It seems that we have reached the decimal's datatype limitations.

As demostrated by the associativity property test, we can get a rounding error when operating with very precise values.

This is expected, as the decimal type does not aims to eliminate the need for rounding but to minimize them.

In our case however, we dont want to implement a rounding policy (at least not at this stage) because we dont know in which currency the client code can be operating with.

So, in order to support every scenario, we find ourselves in the need to use a more precise datatype to handle the amount values.